Refactoring
Removed several unused classes as well as an entire unused class from
the API.
