+++ src/main/java/com/infinityraider/agricraft/apiimpl/FertilizerRegistry.java
@@ -2,7 +2,7 @@
  */
 package com.infinityraider.agricraft.apiimpl;
 
-import com.infinityraider.agricraft.vanilla.BonemealWrapper;
+import com.infinityraider.agricraft.compat.vanilla.BonemealWrapper;
 import com.infinityraider.agricraft.api.fertilizer.IAgriFertilizer;
 import com.infinityraider.agricraft.api.adapter.IAgriAdapterRegistry;
